+++ src/main/java/com/gdschongik/gdsc/domain/discord/listener/JoinCommandListener.java
@@ -1,24 +1,24 @@
-package com.gdschongik.gdsc.global.discord.listener;
+package com.gdschongik.gdsc.domain.discord.listener;
 
 import static com.gdschongik.gdsc.global.common.constant.DiscordConstant.*;
 
-import com.gdschongik.gdsc.domain.discord.handler.IssuingCodeCommandHandler;
+import com.gdschongik.gdsc.domain.discord.handler.JoinCommandHandler;
 import com.gdschongik.gdsc.global.discord.Listener;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import org.jetbrains.annotations.NotNull;
 
 @Listener
 @RequiredArgsConstructor
-public class IssuingCodeCommandListener extends ListenerAdapter {
+public class JoinCommandListener extends ListenerAdapter {
 
-    private final IssuingCodeCommandHandler issuingCodeCommandHandler;
+    private final JoinCommandHandler joinCommandHandler;
 
     @Override
     public void onSlashCommandInteraction(@NotNull SlashCommandInteractionEvent event) {
-        if (event.getName().equals(COMMAND_NAME_ISSUING_CODE)) {
-            issuingCodeCommandHandler.delegate(event);
+        if (event.getName().equals(COMMAND_NAME_JOIN)) {
+            joinCommandHandler.delegate(event);
         }
     }
 }
