+++ src/main/java/com/gdschongik/gdsc/domain/discord/listener/NicknameModifyListener.java
@@ -1,4 +1,4 @@
-package com.gdschongik.gdsc.global.discord.listener;
+package com.gdschongik.gdsc.domain.discord.listener;
 
 import com.gdschongik.gdsc.domain.discord.handler.NicknameModifyHandler;
 import com.gdschongik.gdsc.global.discord.Listener;
